Modularize quiz section
